+++ slick_reporting/generator.py
@@ -5,7 +5,7 @@ import logging
 from inspect import isclass
 
 from django.core.exceptions import ImproperlyConfigured, FieldDoesNotExist
-from django.db.models import Q
+from django.db.models import Q, ForeignKey
 
 from .fields import SlickReportField
 from .helpers import get_field_from_query_text
@@ -112,7 +112,7 @@ class ReportGenerator(object):
                  crosstab_model=None, crosstab_columns=None, crosstab_ids=None, crosstab_compute_reminder=None,
                  swap_sign=False, show_empty_records=None,
                  print_flag=False,
-                 doc_type_plus_list=None, doc_type_minus_list=None, limit_records=False, ):
+                 doc_type_plus_list=None, doc_type_minus_list=None, limit_records=False, format_row_func=None):
         """
 
         :param report_model: Main model containing the data
@@ -162,6 +162,8 @@ class ReportGenerator(object):
         self.crosstab_columns = crosstab_columns or self.crosstab_columns or []
         self.crosstab_ids = self.crosstab_ids or crosstab_ids or []
         self.crosstab_compute_reminder = self.crosstab_compute_reminder if crosstab_compute_reminder is None else crosstab_compute_reminder
+
+        self.format_row = format_row_func or self._default_format_row
 
         main_queryset = main_queryset or self.report_model.objects
         main_queryset = main_queryset.order_by()
@@ -227,8 +229,12 @@ class ReportGenerator(object):
 
             else:
                 self.main_queryset = self._apply_queryset_options(main_queryset)
-                ids = self.main_queryset.values_list(self.group_by_field.attname).distinct()
-                self.main_queryset = self.group_by_field.related_model.objects.filter(pk__in=ids).values()
+
+                if type(self.group_by_field) is ForeignKey:
+                    ids = self.main_queryset.values_list(self.group_by_field.attname).distinct()
+                    self.main_queryset = self.group_by_field.related_model.objects.filter(pk__in=ids).values()
+                else:
+                    self.main_queryset = self.main_queryset.distinct().values(self.group_by_field.attname)
         else:
             self.main_queryset = self._apply_queryset_options(main_queryset, self.get_database_columns())
 
@@ -366,8 +372,17 @@ class ReportGenerator(object):
         )
 
         get_record_data = self._get_record_data
-        data = [get_record_data(obj, all_columns) for obj in main_queryset]
+        format_row = self.format_row
+        data = [format_row(get_record_data(obj, all_columns)) for obj in main_queryset]
         return data
+
+    def _default_format_row(self, row_obj):
+        """
+        Hook where you can format row values like properly format a date
+        :param row_obj:
+        :return:
+        """
+        return row_obj
 
     @classmethod
     def check_columns(cls, columns, group_by, report_model, ):
@@ -381,10 +396,17 @@ class ReportGenerator(object):
         group_by_model = None
         if group_by:
             group_by_field = [x for x in report_model._meta.fields if x.name == group_by][0]
-            group_by_model = group_by_field.related_model
+            if group_by_field.is_relation:
+                group_by_model = group_by_field.related_model
+            else:
+                group_by_model = report_model
 
         parsed_columns = []
         for col in columns:
+            if col in ['__time_series__', '__crosstab__']:
+                #     These are placeholder not real computation field
+                continue
+
             magic_field_class = None
             attr = None
 
@@ -460,7 +482,7 @@ class ReportGenerator(object):
             time_series_columns = self.get_time_series_parsed_columns()
             try:
                 index = self.columns.index('__time_series__')
-                columns[index] = time_series_columns
+                columns[index:index] = time_series_columns
             except ValueError:
                 columns += time_series_columns
 
@@ -469,7 +491,7 @@ class ReportGenerator(object):
 
             try:
                 index = self.columns.index('__crosstab__')
-                columns[index] = crosstab_columns
+                columns[index:index] = crosstab_columns
             except ValueError:
                 columns += crosstab_columns
 
@@ -496,8 +518,8 @@ class ReportGenerator(object):
                     magic_field_class = col
 
                 _values.append({
-                    'name': col + 'TS' + dt[1].strftime('%Y%m%d'),
-                    'original_name': col,
+                    'name': magic_field_class.name + 'TS' + dt[1].strftime('%Y%m%d'),
+                    'original_name': magic_field_class.name,
                     'verbose_name': self.get_time_series_field_verbose_name(magic_field_class, dt),
                     'ref': magic_field_class,
                     'start_date': dt[0],
@@ -548,7 +570,7 @@ class ReportGenerator(object):
             elif series == 'custom':
                 return self.get_custom_time_series_dates()
             else:
-                raise NotImplementedError()
+                raise NotImplementedError(f'"{series}" is not implemented for time_series_pattern')
 
             done = False
             start_date = self.start_date
@@ -581,8 +603,8 @@ class ReportGenerator(object):
                     magic_field_class = col
 
                 output_cols.append({
-                    'name': f'{col}CT{id}',
-                    'original_name': col,
+                    'name': f'{magic_field_class.name}CT{id}',
+                    'original_name': magic_field_class.name,
                     'verbose_name': self.get_crosstab_field_verbose_name(magic_field_class, self.crosstab_model, id),
                     'ref': magic_field_class,
                     'id': id,